pyproject.toml
[project]
name = "vino-students"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "chromadb>=1.0.4",
    "google-genai>=1.10.0",
    "google-generativeai>=0.8.4",
    "langchain>=0.3.23",
    "langchain-core>=0.3.51",
    "langchain-google-genai>=2.0.10",
    "langchain-text-splitters>=0.3.8",
    "pydantic>=2.11.3",
    "pypdf>=5.4.0",
    "pypdf2>=3.0.1",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
    "reflex>=0.7.6",
]
